Address PR review feedback for slide styles
