+++ Docs/Plans/2026-03-17-pr-909-review-feedback-plan.md
@@ -1,0 +1,23 @@
+## Stage 1: Verify PR Feedback
+**Goal**: Confirm each open PR comment against the actual implementation before changing code.
+**Success Criteria**: The backend validation concern and the Presentation Studio maintainability comment are either verified or rejected with technical justification.
+**Tests**: Review PR threads, inspect affected code paths, identify regression coverage needed.
+**Status**: Complete
+
+## Stage 2: Fix Verified Backend Validation Gap
+**Goal**: Reject invalid `appearance_defaults.custom_css` values from visual styles and guard presentation application paths.
+**Success Criteria**: Invalid non-string `custom_css` values are rejected with a controlled `422` response and cannot reach presentation persistence.
+**Tests**: Targeted `pytest` coverage for style creation validation and defense-in-depth when a malformed style payload is resolved.
+**Status**: Complete
+
+## Stage 3: Extract Visual Style Manager UI
+**Goal**: Move custom visual-style editor state, CRUD handlers, and JSX out of `PresentationStudioPage`.
+**Success Criteria**: `PresentationStudioPage` delegates custom-style management to a dedicated component without changing user-visible behavior.
+**Tests**: Existing Presentation Studio page tests remain green; add or adjust targeted tests only if the extraction changes behavior.
+**Status**: Complete
+
+## Stage 4: Verify, Publish, And Resolve Threads
+**Goal**: Re-run backend/UI verification, push the fixes, and respond to the PR comments.
+**Success Criteria**: Relevant tests, typecheck, and Bandit pass for the touched scope; PR threads receive concrete replies tied to the fix.
+**Tests**: Targeted `pytest`, targeted `vitest`, UI `tsc`, Bandit on touched backend files.
+**Status**: In Progress
